Put each reducer in a sepatate file
