+++ lib/reducers/auth/store_auth_step.dart
@@ -1,0 +1,9 @@
+import 'package:adventures_in_tech_world/actions/auth/store_auth_step.dart';
+import 'package:adventures_in_tech_world/models/app/app_state.dart';
+import 'package:redux/redux.dart';
+
+class StoreAuthStepReducer extends TypedReducer<AppState, StoreAuthStep> {
+  StoreAuthStepReducer()
+      : super(
+            (state, action) => state.rebuild((b) => b..authStep = action.step));
+}
