+++ lib/reducers/problems/add_problem.dart
@@ -1,0 +1,9 @@
+import 'package:adventures_in_tech_world/actions/problems/add_problem.dart';
+import 'package:adventures_in_tech_world/models/app/app_state.dart';
+import 'package:redux/redux.dart';
+
+class AddProblemReducer extends TypedReducer<AppState, AddProblem> {
+  AddProblemReducer()
+      : super((state, action) =>
+            state.rebuild((b) => b..problems.add(action.problem)));
+}
