+++ lib/reducers/problems/remove_problem.dart
@@ -1,0 +1,10 @@
+import 'package:adventures_in_tech_world/actions/problems/remove_problem.dart';
+import 'package:adventures_in_tech_world/models/app/app_state.dart';
+import 'package:redux/redux.dart';
+
+class RemoveProblemReducer extends TypedReducer<AppState, RemoveProblem> {
+  RemoveProblemReducer()
+      : super((state, action) => state.rebuild((b) => b
+          ..problems.remove(action.problem)
+          ..displayingProblem = false));
+}
